Extract DynamoDB test setup into a dedicated module
